bin: introduced the bin directory and added the initial CLI build along with README
This introduces the new molt-lms-cli binary so that it is available for customers to easily access. This is currently distributed as a binary since the source code is currently coupled to other packages in another private repository. The plan here is to decouple the CLI and bring it over to this repository once that work is completed. For now, there will be automation to sync changes to the CLI

Added in multiple architecture/OS combination's Go builds so that it supports the binaries needed.

Release Note: None