.github/workflows: automate the docs creation ticket after release
Gets the log between the head of the branch and the latest release tag. Then creates a ticket with the relevant fields, using the official JIRA GH action helper.

Resolves: CC-26297, CC-26370
Release note: Release notes for MOLT features are now more structured and describe customer-facing implications more clearly. To track the work of adding these notes to public documentation, the notes are picked up by a GitHub automation that creates a JIRA ticket corresponding to each MOLT release.