fetch,verify: output result with FmtExport
Previously, with output being printed in FmtBareString, additional single quotes
will be added if the output contains special character, which leads to
inconsistency in the target from the source. This commit is to fix it.

Release Note: bug fix for printing the the output with special characters in the correct format.