verify: support TINYINT to BOOL  mapping
Previously, there were many warnings about mismatching column definitions because there is not a direct translation from a MySQL TINYINT(1), which is used for booleans, and the CRDB BOOL value. This change aims to add support for this in comparable types.

Release Note: TINYINT(1) in MySQL now maps to BOOL types in Cockroach, which resolves a mismatching table definition error.