molt: added in logger field for type to differentiate between task, summary, and data logging (#99)
Added in logger types and Zerolog log context so that logs can be tagged with different contextual information. Also added utility methods that allow a user to retrieve a summary or data logger from their base logger.

Added in utility to better format durations so that they show up in hours seconds minutes format.

Resolves: CC-26366
Release Note: Added tagging for summary and data logs so that customers can filter by "summary" or "data" in a variety of log processing tools and even grep in the command line. This solves the pain point of all the logs going to the same location without any differentiation. Now it's possible to view types of logs one at a time. Summary logs are defined as ones that report aggregated results on either a task or step in a task (i.e. task took 10 minutes, 500 tables processed overall). Data logs are ones that report on individual or row level operations and quantitative markers (i.e. 1000 rows completed so far; failed to verify item with PK 1000)