verify: updated reporting to use data and summary loggers in correct places
Verify centralizes logging in a LogReporter struct. Currently, they all log to the same task logger, but the change here is to get the appropriate loggers and annotate the logs as they come out.

Updated the logging messages to snake case types so that it's easier to process downstream for log processing tools without regexes.

Release Note: Added logging annotations for data level errors / information and summary status logs. In this context, summary logs are ones that are reported about the overall functioning of the job (i.e. 10000 truth rows seen, 5000 mismatches, etc.). The data ones report the exact issues that surfaced during the verify process. These can be filtered so that exact tables + rows can be triaged in one place.