verify: refactored the stats and inconsistency reporting so it's in exported from the inconsistency package (#100)
* verify: updated reporting to use data and summary loggers in correct places

Verify centralizes logging in a LogReporter struct. Currently, they all log to the same task logger, but the change here is to get the appropriate loggers and annotate the logs as they come out.

Updated the logging messages to snake case types so that it's easier to process downstream for log processing tools without regexes.

Release Note: Added logging annotations for data level errors / information and summary status logs. In this context, summary logs are ones that are reported about the overall functioning of the job (i.e. 10000 truth rows seen, 5000 mismatches, etc.). The data ones report the exact issues that surfaced during the verify process. These can be filtered so that exact tables + rows can be triaged in one place.

* verify: refactored the stats and inconsistency reporting so it's in exported from the inconsistency package

Made the summary stats report the stats in the log as individual fields in JSON so that they can more easily be filtered and extracted. Previously, this would require regex extraction, but now can be indexed directly.

Release Note: Changed the output for the running summary statistics so that they are now more easily indexable by log processing tools. Previously, it was more human readable, but fields were not indexable easily out of the box. Now, there is a balance between readability and downstream processing.