verify: made a helper to check dependency of flags on a root flag
Previously, flags were grouped together by the MarkFlagRequiredTogether method that Cobra supplies. This ends up having the downside that all flags marked together must be set, which makes for a worse UX because people need to supply more flags instead of relying on sane defaults.

The change here is to introduce a concept of dependencies so that a flag can mark a one-directional dependency to another flag so that if a child is set without the parent dependency, it will error out. But if the parent is set, the child doesn't necessarily need to be set.

Release Note: Release Note: Marked the "live" and "continuous" flags as dependencies for other sub-configuration flags. Disregard the MarkFlagRequiredTogether note from above. This provides a clearer user experience so that if a flag is set without its dependency, molt will let you know which dependency flag is not correctly set.