verify: handle case where row data comparison cannot be done because of heterogeneous columns
Release Note: Previously, if a comparison between heterogeneous column types occurred, verify would panic and stop execution. With this change to use CompareError instead of Compare, verify no longer panics, but rather records the error and reports it to the logging and also marks the row as mismatching columns. The mismatching columns error message is logged at most once per shard to avoid log spewing. Additionally, the concept of conditional successes is introduced which instructs the user to check all warnings and verify the data. A conditional success means that all comparable row columns were compared and succeeded with the same value.