fetchmetrics: added all metric types and metric initializers
Previously, we only tracked metrics for the number of imported rows. This change aims to add details about progress, durations, errors/exceptions encountered, and the number of tables involved. The metrics will be plumbed through the codebase and capture meaningful data that will allow folks to understand the performance and stability of the system.

Resolves: CC-26354
Release Note: None