fetchmetrics: added all metric types and metric initializers (#111)
* fetchmetrics: added all metric types and metric initializers

Previously, we only tracked metrics for the number of imported rows. This change aims to add details about progress, durations, errors/exceptions encountered, and the number of tables involved. The metrics will be plumbed through the codebase and capture meaningful data that will allow folks to understand the performance and stability of the system.

Resolves: CC-26354
Release Note: None

* fetch: added in resource key method and ability to track the progression of COPY by row count

Release Note: None