fetch: Add row count per file to blob resource
This commit adds a row count field to ever storage resource. The
reason for this is mainly for progress tracking but more specfically
for the IMPORT use case where we are not able to know the number of rows
imported. We now have this information since each resource has its own
file count.

Release note: None