fetch: fix the deadlock in csv pipe
Previous, if error happens when writing result to the storage (gcp/aws/local),
the error won't be properly propagated back to the main routine in exportTable(),
causing the main routine to hang.

This patch is to fix it so that the main process exits in this case.

Release note (bug fix): resolve deadlock when failed to upload csv to the data
blob storage.