fetch: added an exception reporter helper and integrated in copy/direct copy/import (#120)
Now there is a helper that parses the PG error, extracts key information necessary to store in the exception log, and writes it / links it to the fetch run. Also added in portions of the code to create the dependency tables and status runs so that the relations are correct.

Refactored fetch so that all the init happens in the fetch method instead of the cmd/fetch entrypoint. This makes the exception logging setup simpler.

Added better test coverage so that the helpers and the datadriven e2e portions are covered.

Added the stage field to better determine the stage at which an exception happened. Removed unnecessary fields for status: finished_at and status.

Resolves: CC-26806
Release Note: None