fetch: fix csv pipe deadlock + add row count per file to blob resource (#125)
* fetch: isolate the GenerateDatastore helper func

So that it can be reused in the test.

Release note: None

* fetch: isolate the GetLocalStoreAddrs helper func

So that it can be reused in the test.

Release note: None

* datablobstorage: put mock functions in testutil.go

Just a simple movement of codes.

Release note: None

* fetch: Add row count per file to blob resource

This commit adds a row count field to ever storage resource. The
reason for this is mainly for progress tracking but more specfically
for the IMPORT use case where we are not able to know the number of rows
imported. We now have this information since each resource has its own
file count.

Release note: None

* fetch: fix the deadlock in csv pipe

Previous, if error happens when writing result to the storage (gcp/aws/local),
the error won't be properly propagated back to the main routine in exportTable(),
causing the main routine to hang.

This patch is to fix it so that the main process exits in this case.

Release note (bug fix): resolve deadlock when failed to upload csv to the data
blob storage.

* fetch: adding test for failed export

Release note: None

---------

Co-authored-by: Jeremy Yang <jyang@cockroachlabs.com>