cmd/fetch: Implement schema creation entrypoint in cli (#128)
* cmd/fetch: add --table-handling flag

This serves as an entrypoint for schema creation, and replaces the existing
--truncate flag.

Release note (cli): introduce the --table-handling flag and deprecate the --truncate
flag for molt fetch. The implementation is not ready yet and the usage this flag is
temporarily gated.

* fetch: enable truncate table option for all fetch mode

Previously, the truncate option is ignored if the fetch is under direct-copy mode.
This patch is to fix this.

Release note (bug fix): enable truncate table option for all fetch mode.