fetch: Fix csv write deadlock
This commit fixes a csv write deadlock issue that was caused by
the adding of numRows causing writes to be blocked due to how
certain goroutines were being scheduled. The write/read threads
would be in gopark state and unable to proceed.

This changes takes inspiration from the http2 package in memory
pipe with sync.cond to ensure that there are signals being sent
to either end of the pipe.

Release note: Fixes csv write deadlock on large file exports