fetch: fix ordering of primary key
Previously, in cases where a table in fetch/verify has a compund primary key
whose ordering doesn't match that of the order of the columns in the base table,
fetch and verify would order the columns incorrectly. This would produce
incorrect results in verify, but more importantly, it would make fetch insanely
slow. The reason why has to do with how fetch runs export. To avoid issuing a
single large query, fetch breaks the export down into batch-sized chunks (the
default batch size is 100K rows). To ensure that it knows where it stopped, it
leverages the primary key and issues a SELECT <column names> order by <PK
columns> limit <batch_size>. Then, it records the PK value at which it left off
and issues subsequent SELECT <column names> where PK > <left off value> order
by <PK columns> limit <batch_size>.

The problem with not getting the PK ordering right is that instead of leveraging
the PK for the query, fetch instead has to sort the entire table each time it
issues a query. This is most pronounced in cases where the table being fetched
is large. For a 104M table, tests showed that it would take 17 hours to perform
the export phase. With the fix to get the PK ordering correct, the export takes
3 minutes.