fetch: fix bug with float32 type
Previously, to convert a float32 golang value to a tree.DFloat, which is a float64, additional non-zero decimals are padded (e.g. 123.12 -> 123.12000274658203), leading to incorrect result when comparing the 2 datums. This commit fixes it by making sure the new tree.DFloat is consistent with the original float32 value.

We also fixed a format issue with float type so that when negative float value is exported to a csv, it is not bracket. It's because `(-123,12)` cannot be parsed as float when being imported to a table.

Release note (bug fix): correctly support verification of float32 (mysql's float4) value, and fmt issue with float.