mysqlcov: fix bits conversion (#156)
Previous, the conversion of the bits type value from mysql to crdb is wrong, as it didn't correctly parse the value passed via the []uint8 val variable. This PR is to fix it and add tests for it.

Release note (bug fix): enable conversion of bit type entry from mysql to crdb.