fetch: added better testing for continuation and fixed behavior for no exceptions log found (#157)
Previously, when no exception logs for a given token were found, we would silently operate and not let the user know. Now we error and let the user know that there were no exception logs found, which helps the user to use a valid token. This could lead to less confusing behavior.

This change also adds coverage for the following cases:
- Continuing from a file greater than part 1
- If a continuation token is no longer present, we error and let user know
- If a continuation token links to an exception log that doesn't have a file name, we error out because no file name is present

Resolves: CC-27047
Release Note: None