verify: add explicit uuid <> string, bool <> tinyint conversion
Due to some mismatching type definitions of mysql and cockroachdb, when comparing row values, verify may emit "comparson between x and y types are not supported" error. We've seen this when it tries comparing
- UUID with String
- TINYINT (mysql) with BOOL (crdb)

This commit is to add explicit conversion between these types to enable a comparison.

Informs: #103

Release note: None