fetch: unified continuation mode behavior so that specific token and multiple token behavior is the same
Previously, the multiple token and single-specific token case were treated differently, leading to potential continuation token loss when errors occur during non continuation logic. This means that we previously truncated the exceptions table before we even ran the import. Now, we unify the logic so that the behavior is as follows: if a table associated with a token fails, we update the associated entry with new information; if it succeeds, we remove the token. This happens within goroutines.

We also added new test cases and a mutex to prevent the `conn busy` pgx issue that happens right now because pgx.Conn is not threadsafe.

Release Note: None