fetch: use a cloned connection for exceptions log reporting
Previously, Jane highlighted that if a connection dies while processing a copy or import, the exception log cannot be successfully written if the same conneciton is used. In order to avoid this case, we create a clone of this connection before we use it to record the exception.

Release Note: None