fetch: add tests for sharding export
We added tests to test the case where error happens in
1. Exporting data from the sql source to the pipe;
2. Init new csv writer for a new iteration;
3. Write the data to the csv writer.

The actual code change is minimal -- we just ensure the goroutines for all shards share the context. It means if one of the shard failed in exporting, the other shards will exit their own goroutines immediately as well.

Release note: None