verfiy: ignore dropped columns (#189)
This commit adds a attisdropped clause for pg/crdb queries when
getting columnns to verify. Previously, this would return and
when running LoadDataType in the PGX driver, it would cause an
issue as the dropped column could not be resolved. We now omit
dropped columns from being returned since we are not verifying them
anyways.

Release Note: Fixes a bug for pg/crdb verify that causes an error
when there are dropped columns on the table